Fix genesis block creating bug
The genesis block should be created from the genesis batch if
the node index=0 and if the sawtooth directory does not exist
or if the diretory exists but the block-chain-id file does not
exist. The genesis.batch file is deleted and the block-chain-id
file is create when the genesis block is created.